requirements.txt
streamlit
pandas
numpy
scikit-learn
xgboost
python-docx
joblib
matplotlib